requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
djangorestframework>=3.14
psycopg2-binary==2.9.6